Don't perform check_messages_for_completion
The job was causing DB contention so it was shut off in prod about a
month ago. We forgot about it and never turned it back on. Now is not a
good time to turn it back on, so we'll try really hard to not forget
again. Promise.

(cherry picked from commit 4b3fd00e10183468050615e832f7d285b2ed4d64)